Transition "HWToSV" to "ProceduralCoreToSV" pass.
