+++ include/circt/Conversion/Passes.h
@@ -32,7 +32,6 @@
 #include "circt/Conversion/HWToLLHD.h"
 #include "circt/Conversion/HWToLLVM.h"
 #include "circt/Conversion/HWToSMT.h"
-#include "circt/Conversion/HWToSV.h"
 #include "circt/Conversion/HWToSystemC.h"
 #include "circt/Conversion/HandshakeToDC.h"
 #include "circt/Conversion/HandshakeToHW.h"
@@ -41,6 +40,7 @@
 #include "circt/Conversion/LoopScheduleToCalyx.h"
 #include "circt/Conversion/MooreToCore.h"
 #include "circt/Conversion/PipelineToHW.h"
+#include "circt/Conversion/ProceduralCoreToSV.h"
 #include "circt/Conversion/SCFToCalyx.h"
 #include "circt/Conversion/SMTToZ3LLVM.h"
 #include "circt/Conversion/SeqToSV.h"
