+++ lib/CAPI/Conversion/CMakeLists.txt
@@ -23,13 +23,13 @@ add_circt_public_c_api_library(CIRCTCAPIConversion
   CIRCTHWToLLVM
   CIRCTHWToBTOR2
   CIRCTHWToSMT
-  CIRCTHWToSV
   CIRCTHWToSystemC
   CIRCTLLHDToLLVM
   CIRCTLoopScheduleToCalyx
   CIRCTLTLToCore
   CIRCTMooreToCore
   CIRCTPipelineToHW
+  CIRCTProceduralCoreToSV
   CIRCTSCFToCalyx
   CIRCTSeqToSV
   CIRCTSimToSV
