+++ lib/Conversion/CMakeLists.txt
@@ -18,12 +18,12 @@ add_subdirectory(HWToBTOR2)
 add_subdirectory(HWToLLHD)
 add_subdirectory(HWToLLVM)
 add_subdirectory(HWToSMT)
-add_subdirectory(HWToSV)
 add_subdirectory(HWToSystemC)
 add_subdirectory(LLHDToLLVM)
 add_subdirectory(LoopScheduleToCalyx)
 add_subdirectory(MooreToCore)
 add_subdirectory(PipelineToHW)
+add_subdirectory(ProceduralCoreToSV)
 add_subdirectory(SCFToCalyx)
 add_subdirectory(SeqToSV)
 add_subdirectory(SimToSV)
